apps/analyzer/pyproject.toml
[tool.coverage.run]
branch = true
source = [ "metadata_analyzer" ]

[tool.coverage.report]
exclude_lines = ['if TYPE_CHECKING:']
show_missing = true

[tool.pytest.ini_options]
addopts = "--cov --cov-report html:'../../coverage/apps/analyzer/html' --cov-report xml:'../../coverage/apps/analyzer/coverage.xml' --html='../../reports/apps/analyzer/unittests/html/index.html' --junitxml='../../reports/apps/analyzer/unittests/junit.xml'"

[tool.poetry]
name = "metadata-analyzer"
version = "1.0.0"
description = "Automatically generated by Nx."
authors = [ ]
license = 'Proprietary'
readme = 'README.md'

  [[tool.poetry.packages]]
  include = "metadata_analyzer"

  [tool.poetry.dependencies]
  python = ">=3.9,<3.12"
  flask = "^3.0.3"
  python-dotenv = "^1.0.1"
  sqlalchemy = "^2.0.36"
  psycopg = "^3.2.3"

  [tool.poetry.group.dev.dependencies]
  autopep8 = "2.0.2"
  flake8 = "6.0.0"
  pytest = "7.3.1"
  pytest-sugar = "0.9.7"
  pytest-cov = "4.1.0"
  pytest-html = "3.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
